calgary-urban-intelligence-dashboard/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
requests==2.31.0
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
gunicorn==21.2.0
transformers==4.53.2
huggingface_hub==0.33.4
torch>=2.1.0 